test(parser): normalize parser engine test structure
